ReferencesAnalyzer main class + Entries Separators
Added the ReferencesAnalyzer class. In order to analyze a file or a folder, a ReferencesAnalyzer object must be created with the given path, then the process() method must be called in order to start analysis.
ReferencesYearsInfo - used to create an EnumMap to hold information
The separtor classes are used to separate messed up references paragraphs into different reference entries (paragraph extraction from pdf files was sometimes incorrect).